Relocate inventory storage responsibilities to delegate of GameManager called PlayerInventory
