+++ Assets/Scripts/Inventory/PlayerInventory.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 52126f5ad39f404e95c5d90cc58ab52e
+timeCreated: 1683423338
