+++ Assets/Scripts/Triggers/TriggerOnTagEnter.cs
@@ -29,7 +29,7 @@ namespace Triggers
                 (CustomTag == "" || 
                  other.CompareTag(CustomTag)))
             {
-                if (RequireItem != "" && !GameManager.instance.itemIsPossessed(RequireItem))
+                if (RequireItem != "" && !GameManager.instance.getInventory().IsItemPossessed(RequireItem))
                 {
                     this.keyUI.ShowNextFrame();
                     return;
